Fix timer logic to use Unix time rather than setInterval
The timer was set to render a new circle every 100ms to match the
timer progress, but inconsistencies in intraststem processing times
as well as throttling from the displayState of the window (the latter
being the biggest factor) make the timer inaccurate. Basing it off of
Unix time makes the timer display render at a speed up to every 100ms
while still being sensitive to slower speeds.