Add logic to group cycles into 4 chunks
The Pomodoro techinque involves grouping sessions into 4 cycles where
a longer break is triggered at the end. This is now worked into the logic.
Encouragement messages are incorporated and reset button as well.

The instructions shown upon app load were brief for simplicity's sake,
so on click they now lead to an expanded description of the Pomodoro
technique.